The first division
Write a program that reads 10 numbers from the input and at the end prints the number that has the highest number of divisors of the first number along with the number of its first divisors in the output. If there are several numbers in this state, print the largest one.