Add files via upload
Iran, Portugal, Spain and Morocco are present in Group B of the World Cup. Write a program that, upon receiving the results of the games, will print the team name, the number of wins and losses, and the difference in goals and points respectively in one line. Each team should be printed in order of points in one line. (If the points are equal, the number of wins will be considered. If both the number of wins and the points are equal, they will be printed in alphabetical order.)

Note: The team gets zero points in case of loss, one point in case of draw and three points in case of win.
Goal difference is the difference between goals scored and goals conceded by a team

Read the results of the games in the following order: (in the sample input, the left number corresponds to the right team.)
Iran - Spain
Iran - Portugal
Iran - Morocco
Spain - Portugal
Spain - Morocco
Portugal - Morocco