Poll of favorite genres
In a survey, moviegoers were asked to list 3 of their favorite genres. They are given 6 different genres to choose from, including:
Horror, Romance, Comedy, History, Adventure, Action

Write a program that takes the number of people, then takes the name of each person with their favorite genres and prints the name of each genre and the number of people interested in that genre in the order of the most interested in the output. (If the level of interest in different genres is the same, print in the order of the English alphabet in the output.) If no genre is selected, consider its value as zero and print the name and number 0 in the output.