Index words
Write a program that prints index words (words that start with capital letters) along with the word number (the most common word) from a text. If a word with this feature is not found in the text, it will print None in the output. You should not consider the words at the beginning of the sentence as index words. (number the words starting from one)

Numbers are not counted except index words. The only sign used in the sentence is the comma. Be sure to remove the dot or comma at the end of the word.