current age
Write a program that gives the age of a person using the datetime library class in Python. If the input date is not correct, print the word WRONG in the output. Date format is yyyy/mm/dd. (For example, if it enters 15 for the month, because we only have 12 months, it should print WRONG in the output. The same applies to the number of days, and the maximum is 31)

Note: The input and output of the following example is related to the execution of the program on 02/01/2019.