Employee information
A database containing the name, height, weight of the employees of a company is designed. Write a program that reads the information from this database file and prints the employees in order of height along with their name and weight. (If they were the same height, print the person with lower weight higher) It is guaranteed that no two people are the same height and weight.

The information format of the database table is as follows: