Correct email
Write a program that prints OK if the email entered by the user matches the following pattern, and prints WRONG if it does not match the pattern. (The output must be in capital letters.)
The pattern is such that it contains two strings and an expression expression (including number and string, the presence of number and string is not required at the same time and can only include string) which is represented by @ and . are separated, it is considered as correct email.